Move material on source positions to a new module, Position
